Move exported code to pkg
